* lazy initialization for regexp * fix unit tests * fix ConvertIntervalYM_DTY with wrong unsigned integer usage * code formatting

+++ v2/converters/other_types.go
@@ -42,12 +42,6 @@ func ConvertBinaryFloat(bytes []byte) float64 {
 	}
 	exp := len(strX) - index - 1
 	return math.Floor(float64(x)*math.Pow10(exp)) / math.Pow10(exp)
-	//test2 := float64(test)
-	//return test2
-	//if u > (1 << 31) {
-	//	return -math.Float32frombits(u)
-	//}
-	//return math.Float32frombits(^u)
 }
 
 func ConvertBinaryDouble(bytes []byte) float64 {
@@ -65,15 +59,6 @@ func ConvertBinaryDouble(bytes []byte) float64 {
 	}
 	u := binary.BigEndian.Uint64(bytes)
 	return math.Float64frombits(u)
-	//if neg {
-	//	out = -out
-	//}
-	//return out
-	//if u > (1 << 63) {
-	//	out = -out
-	//}
-	//return math.Float64frombits(u)
-	//return math.Float64frombits(^u)
 }
 
 /*
@@ -87,7 +72,7 @@ func ConvertIntervalYM_DTY(val []byte) string {
 	   months + 60
 	*/
 	uyears := binary.BigEndian.Uint32(val)
-	years := int(uyears - uint32(2147483648))
+	years := int(uyears) - 2147483648
 	if years >= 0 {
 		months := val[4] - 60
 		return fmt.Sprintf("+%02d-%02d", years, months)
@@ -108,13 +93,13 @@ func ConvertIntervalDS_DTY(val []byte) string {
 	   nanoseconds
 	*/
 	udays := binary.BigEndian.Uint32(val)
-	days := int(udays - uint32(2147483648))
+	days := int(udays) - 2147483648
 	if days >= 0 {
 		hours := val[4] - 60
 		mins := val[5] - 60
 		secs := val[6] - 60
 		uns := binary.BigEndian.Uint32(val[7:])
-		ns := (int(uns - uint32(2147483648))) / 1000
+		ns := (int(uns) - 2147483648) / 1000
 		return fmt.Sprintf("+%02d %02d:%02d:%02d.%06d", days, hours, mins, secs, ns)
 	}
 	days = -days
@@ -122,6 +107,6 @@ func ConvertIntervalDS_DTY(val []byte) string {
 	mins := 60 - val[5]
 	secs := 60 - val[6]
 	uns := binary.BigEndian.Uint32(val[7:])
-	ns := -(int(uns - uint32(2147483648))) / 1000
+	ns := -(int(uns) - 2147483648) / 1000
 	return fmt.Sprintf("-%02d %02d:%02d:%02d.%06d", days, hours, mins, secs, ns)
 }
